Track value uses with linked lists
Using that hash set is not as ideal as I'd hoped, since we can't add/remove uses during enumeration. Now, each instruction keep an additional array of "Use" nodes, which makes it possible to do O(1) removes and replaces.

The Users() enumerator does not prevent duplicates, so if a instruction uses a value twice, it will be enumerated twice.

There's probably a better way to encode uses, but i'm just copying llvm here.