Refactor and optimize custom sets
As part of my procrastination routine, I noticed that RefSet was slower than HashSet because of array covariance. A workaround for this issue involves an unsafe reinterpret cast, and it gives back the 30-40% perf improvement over HashSet with non-sealed types.

The BitSet changes are half cosmetic and half micro opts.