Fix importing of leave instructions
Regions are represented implicitly in the CFG using guard/leave instructions. Prior to this commit, nesting depth was not considered when importing leave instructions, which resulted in malformed/ambiguous IR.

The region tree class is perhaps slightly overkill, since we could in theory directly calculate the nesting depth using only the region clauses, but at least it avoids issues with ordering, and it's easier to reason about.

Contributes to #6