Method bisecting cli switch
This is very similar to `git bisect`, but the entire state is supplied via the switch.

DistIL.Cli.exe ... --bisect gbg

Each character indicate whether the last compilation contained the bug (`b`ad) or not (`g`ood).
If used, a log containing all compiled methods is written to the working directory.