Introduce SelectInst and basic support for if-conversion
The conversion pass only handles diamond-like sub graphs, where
exactly two branch blocks merge at another starting with a phi.

It is limited to a single phi and 2 non-side-effecting instructions per branch.

The conversion pass will also simplify some conditionals like
"if (A) if (B)" to "A & B", which may help with transforms.