Fix method DFS traversal helper
The old algorithm was recently introduced in 26b6d0ea. Although it is simpler, it
may produce incorrect post-orderings, which may become problematic in the future.

Coincidentally it was exactly the same algorithm used in RyuJIT, and the bug
is so subtle that they just recently discovered it :') - dotnet/runtime#82950

This also adds a few micro opts in SuccIterator for perf and mem overhead.