Implement a crude GVN pass
There are no heuristics yet, so this might also introduce some spills on
the final asm. JIT already does some VN, so we may want to tune it down
for most simple expressions - the real goodness is probably going to be
in the call and load/store stuff anyway...

Load/store and call clobbering is tracked using a simple backward scan,
which is surprisingly fast for normal programs but will surely scale very
horribly. LLVM seems to be doing a similar scan for its MemDepAnalysis,
so it's probably okay enough.

Closes #9