Temporarily disable broken AssertProp pass
ILSpy crashes because of it. Apparently related to changes in bca2d64,
but I already wasted enough time trying to find the bug.

The intent behind this pass is in enabling folding of branches that can
be statically determined (like `is T[]` and etc), which are common after
inlining Linq methods, but that is overall not really meaningful yet.